refactor(core): establish P1 decomposition guardrails
Move the core decomposition plan into docs/plans and add the architecture
guardrail doc for future bitfun-core split work.

Introduce the product-full safety net without reducing the default core
capability set, and make current product crates opt into the full runtime
explicitly.

Add bitfun-core-types for shared AI error DTOs while preserving the existing
bitfun-events re-export path. Move the existing terminal-core and tool-runtime
packages to workspace-level crate paths without changing their package or lib
names.

Update AGENTS, README, and CONTRIBUTING guidance so future core decomposition
work follows the guardrails and uses the new support crate locations.
