+++ src/crates/core/AGENTS.md
@@ -28,6 +28,13 @@ SessionManager → Session → DialogTurn → ModelRound
 - Avoid host-specific APIs such as `tauri::AppHandle`
 - Use shared abstractions such as `bitfun_events::EventEmitter`
 - Desktop-only integrations belong in `src/apps/desktop`, then flow through transport/API layers
+- During core decomposition, `bitfun-core` is a compatibility facade and full
+  product runtime assembly point. New modules should prefer the extracted owner
+  crate listed in `docs/architecture/core-decomposition.md`.
+- Do not add new cross-layer references from `service` to `agentic` without a
+  small port/interface boundary.
+- Do not move platform-specific logic, build-script behavior, or product
+  capability selection into shared core as part of decomposition.
 
 Narrower rules already exist:
