+++ src/crates/events/Cargo.toml
@@ -4,6 +4,7 @@ version.workspace = true
 edition.workspace = true
 
 [dependencies]
+bitfun-core-types = { path = "../core-types" }
 async-trait = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
