+++ src/crates/terminal/docs/STREAMING_OUTPUT_COLLECTION.md
