+++ src/crates/terminal/src/api.rs
