+++ src/crates/terminal/src/config/mod.rs
