+++ src/crates/terminal/src/config/types.rs
