+++ src/crates/terminal/src/events.rs
