+++ src/crates/terminal/src/pty/data_bufferer.rs
