+++ src/crates/terminal/src/session/binding.rs
