+++ src/crates/terminal/src/session/manager.rs
