+++ src/crates/terminal/src/session/mod.rs
