+++ src/crates/terminal/src/session/persistent.rs
