+++ src/crates/terminal/src/session/serializer.rs
