+++ src/crates/terminal/src/session/singleton.rs
