+++ src/crates/terminal/src/shell/detection.rs
