+++ src/crates/terminal/src/shell/integration.rs
