+++ src/crates/terminal/src/shell/mod.rs
