+++ src/crates/terminal/src/shell/profiles.rs
