+++ src/crates/terminal/src/shell/scripts/shellIntegration-bash.sh
