+++ src/crates/terminal/src/shell/scripts/shellIntegration-rc.zsh
