+++ src/crates/terminal/src/shell/scripts/shellIntegration.fish
