+++ src/crates/terminal/src/shell/scripts/shellIntegration.ps1
