+++ src/crates/terminal/src/shell/scripts_manager.rs
