+++ src/crates/tool-runtime/src/util/string.rs
